Rename package, upgrade dependencies and correct syntax to match
I've forked this into a separate package in order to make it work with
React 17. In particular, the enzyme-adapter-react-15 prevents it from
installing at all in a React 17 project:

    $ npm i -D chai-enzyme
    npm ERR! code ERESOLVE
    npm ERR! ERESOLVE unable to resolve dependency tree
    npm ERR!
    npm ERR! While resolving: customer-data-platform-client@1.0.0
    npm ERR! Found: react@17.0.2
    npm ERR! node_modules/react
    npm ERR!   react@"^17.0.2" from the root project
    npm ERR!
    npm ERR! Could not resolve dependency:
    npm ERR! peer react@"^0.14.0 || ^15.0.0-0 || ^16.0.0-0" from chai-enzyme@1.0.0-beta.1
    npm ERR! node_modules/chai-enzyme
    npm ERR!   dev chai-enzyme@"*" from the root project
    npm ERR!
    npm ERR! Fix the upstream dependency conflict, or retry
    npm ERR! this command with --force, or --legacy-peer-deps
    npm ERR! to accept an incorrect (and potentially broken) dependency resolution.

As a first step I'm renaming the package into my private scope and
editing the doc to reflect this. While I'm at it I figured I'd update
all the packages to get up to date with everything.